Very initial draft POC.
