+++ lib/shopify-cli/commands/deploy.rb
@@ -4,34 +4,14 @@ require 'shopify_cli'
 
 module ShopifyCli
   module Commands
-    class Deploy < ShopifyCli::Command
-      autoload :Heroku, 'shopify-cli/commands/deploy/heroku'
-
-      def call(args, _name)
-        subcommand = args.shift
-        case subcommand
-        when 'heroku'
-          Heroku.call(@ctx, args)
-        else
-          @ctx.puts(self.class.help)
+    class Deploy < ShopifyCli::ContextualCommand
+      class << self
+        def register_contextual_command
+          project_type = Project.current.config['project_type']
+          if project_type == :app || :script
+            register_for_context 'deploy', project_type, 'shopify-cli/commands/deploy'
+          end
         end
-      end
-
-      def self.help
-        <<~HELP
-          Deploy the current app project to a hosting service. Heroku ({{underline:https://www.heroku.com}}) is currently the only option, but more will be added in the future.
-            Usage: {{command:#{ShopifyCli::TOOL_NAME} deploy [heroku]}}
-        HELP
-      end
-
-      def self.extended_help
-        <<~HELP
-          Subcommands:
-
-          * heroku: Deploys the current app project to Heroku. Usage:
-
-              {{command:#{ShopifyCli::TOOL_NAME} deploy heroku}}
-        HELP
       end
     end
   end
