+++ lib/shopify-cli/entry_point.rb
@@ -8,9 +8,15 @@ module ShopifyCli
       def call(args, ctx = Context.new)
         task_registry = ShopifyCli::Tasks::Registry
 
+        orig_args = args.dup
         before_resolve(args)
-
         command, command_name, args = ShopifyCli::Resolver.call(args)
+
+        while command.has_more_context?
+          command.register_contextual_command
+          command, command_name, args = ShopifyCli::Resolver.call(orig_args)
+        end
+
         executor = ShopifyCli::Executor.new(ctx, task_registry, log_file: ShopifyCli::LOG_FILE)
         ShopifyCli::Monorail.log.invocation(command_name, args) do
           executor.call(command, command_name, args)
